requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
et-xmlfile==1.0.1
Flask==0.12.2
Flask-Cors==3.0.3
idna==2.6
itsdangerous==0.24
jdcal==1.3
Jinja2==2.10
MarkupSafe==1.0
olefile==0.44
openpyxl==2.4.9
Pillow==4.3.0
pkg-resources==0.0.0
pymongo==3.5.1
python-binary-memcached==0.26.1
requests==2.18.4
selenium==3.7.0
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
